requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-environ==0.11.2
dj-database-url==2.2.0
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.0.8
gunicorn==23.0.0
whitenoise==6.8.2
packaging==24.2
